feat(custom-chart-web): move utils to shared/charts
