+++ packages/pluggableWidgets/custom-chart-web/src/CustomChart.tsx
@@ -1,9 +1,8 @@
-import { getPlaygroundContext } from "@mendix/shared-charts/main";
+import { constructWrapperStyle, getPlaygroundContext } from "@mendix/shared-charts/main";
 import { createElement, Fragment, ReactElement } from "react";
 import { CustomChartContainerProps } from "../typings/CustomChartProps";
 import { useCustomChart } from "./hooks/useCustomChart";
 import "./ui/CustomChart.scss";
-import { constructWrapperStyle } from "./utils/utils";
 
 const PlaygroundContext = getPlaygroundContext();
