+++ packages/pluggableWidgets/custom-chart-web/src/utils/utils.ts
@@ -1,7 +1,4 @@
 import { Config, Data, Layout } from "plotly.js-dist-min";
-import { CSSProperties } from "react";
-import { CustomChartContainerProps } from "typings/CustomChartProps";
-
 export function parseData(staticData?: string, attributeData?: string, sampleData?: string): Data[] {
     let finalData: Data[] = [];
 
@@ -52,34 +49,4 @@ export function parseConfig(configOptions?: string): Partial<Config> {
         console.error("Error parsing chart config:", error);
         return {};
     }
-}
-
-function getHeightScale(height: number, heightUnit: "pixels" | "percentageOfParent" | "percentageOfView"): string {
-    return `${height}${heightUnit === "pixels" ? "px" : heightUnit === "percentageOfView" ? "vh" : "%"}`;
-}
-
-export function constructWrapperStyle(props: CustomChartContainerProps): CSSProperties {
-    const { widthUnit, heightUnit, minHeightUnit, maxHeightUnit, width, height, minHeight, maxHeight, OverflowY } =
-        props;
-
-    const wrapperStyle: Pick<CSSProperties, "width" | "height" | "minHeight" | "maxHeight" | "maxWidth" | "overflowY"> =
-        {};
-
-    wrapperStyle.width = `${width}${widthUnit === "pixels" ? "px" : "%"}`;
-    if (heightUnit === "percentageOfWidth") {
-        wrapperStyle.height = "auto";
-
-        if (minHeightUnit !== "none") {
-            wrapperStyle.minHeight = getHeightScale(minHeight, minHeightUnit);
-        }
-
-        if (maxHeightUnit !== "none") {
-            wrapperStyle.maxHeight = getHeightScale(maxHeight, maxHeightUnit);
-            wrapperStyle.overflowY = OverflowY;
-        }
-    } else {
-        wrapperStyle.height = getHeightScale(height, heightUnit);
-    }
-
-    return wrapperStyle;
 }
