+++ packages/shared/charts/src/utils/chartStyles.ts
@@ -1,0 +1,32 @@
+import { CSSProperties } from "react";
+import { CustomChartContainerProps } from "../../../../pluggableWidgets/custom-chart-web/typings/CustomChartProps";
+
+function getHeightScale(height: number, heightUnit: "pixels" | "percentageOfParent" | "percentageOfView"): string {
+    return `${height}${heightUnit === "pixels" ? "px" : heightUnit === "percentageOfView" ? "vh" : "%"}`;
+}
+
+export function constructWrapperStyle(props: CustomChartContainerProps): CSSProperties {
+    const { widthUnit, heightUnit, minHeightUnit, maxHeightUnit, width, height, minHeight, maxHeight, OverflowY } =
+        props;
+
+    const wrapperStyle: Pick<CSSProperties, "width" | "height" | "minHeight" | "maxHeight" | "maxWidth" | "overflowY"> =
+        {};
+
+    wrapperStyle.width = `${width}${widthUnit === "pixels" ? "px" : "%"}`;
+    if (heightUnit === "percentageOfWidth") {
+        wrapperStyle.height = "auto";
+
+        if (minHeightUnit !== "none") {
+            wrapperStyle.minHeight = getHeightScale(minHeight, minHeightUnit);
+        }
+
+        if (maxHeightUnit !== "none") {
+            wrapperStyle.maxHeight = getHeightScale(maxHeight, maxHeightUnit);
+            wrapperStyle.overflowY = OverflowY;
+        }
+    } else {
+        wrapperStyle.height = getHeightScale(height, heightUnit);
+    }
+
+    return wrapperStyle;
+}
